+++ artifacts/ArtifactHandling/NuGet/Initialize-NuGetFeeds.ps1
@@ -25,8 +25,9 @@
 
         if ($global:extendedEnv.CustomNugetFeeds) {            
             Write-Host "Collecting custom nuget feeds"
-            [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($global:extendedEnv.CustomNugetFeeds)) | 
-                ConvertFrom-Json |
+            $customNugetFeedsBase64 = $global:extendedEnv.CustomNugetFeeds
+            $customNugetFeedsJson = [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($customNugetFeedsBase64))
+            ($customNugetFeedsJson | ConvertFrom-Json) |
                 Where-Object { $_ } |
                 ForEach-Object {
                     $url = $_.feedUrl